build setup

+++ IQCare Management/IQCare.SCM/frmPatientDrugDispense.cs
@@ -169,7 +169,7 @@ namespace IQCare.SCM
         /// <summary>
         /// The selling price
         /// </summary>
-        private double theSellingPrice = 0;
+        private decimal theSellingPrice = 0;
 
         /// <summary>
         /// The strength
@@ -2019,7 +2019,7 @@ namespace IQCare.SCM
                     theBatchId = Convert.ToInt32(theDV[0]["BatchId"]);
                     txtExpirydate.Text = ((DateTime)theDV[0]["ExpiryDate"]).ToString(GblIQCare.AppDateFormat);
 
-                    theSellingPrice = Convert.ToDouble(theDV[0]["SellingPrice"]);
+                    theSellingPrice = Convert.ToDecimal(theDV[0]["SellingPrice"]);
                     theItemTypeId = Convert.ToInt32(theDV[0]["ItemTypeID"]);
                     this.GetPrice(theItemId, theItemTypeId);
                     theConfigSellingPrice = Convert.ToDecimal(theDV[0]["ConfigSellingPrice"]);
@@ -2579,7 +2579,7 @@ namespace IQCare.SCM
                 txtBatchNo.Text = theDV[0]["BatchNo"].ToString();
                 theBatchId = Convert.ToInt32(theDV[0]["BatchId"]);
                 txtExpirydate.Text = ((DateTime)theDV[0]["ExpiryDate"]).ToString(GblIQCare.AppDateFormat);
-                theSellingPrice = Convert.ToDouble(theDV[0]["SellingPrice"]);
+                theSellingPrice = Convert.ToDecimal(theDV[0]["SellingPrice"]);
                 theItemTypeId = Convert.ToInt32(theDV[0]["ItemTypeID"]);
                 this.GetPrice(theItemId, theItemTypeId);
